Fix: distX & speedX are now floats along with MAX_SPEED

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/Vision.java
@@ -26,7 +26,7 @@ public class Vision extends LinearOpMode {
     private int closestY;
     private int screenWidth = 320;
     private int screenHeight = 240;
-    private int MAX_SPEED = 1;
+    private float MAX_SPEED = 1;
 
     @Override
     public void runOpMode() {
@@ -73,8 +73,8 @@ public class Vision extends LinearOpMode {
                 If the Y values happen to be the same, it will choose the one closest to the center */
             }
 
-                int distX = closestX - (screenWidth/2);
-                int speedX = Math.max((((double) distX/160) * MAX_SPEED), 0.25);
+                float distX = closestX - (screenWidth/2);
+                float speedX = Math.max((((double) distX/160) * MAX_SPEED), 0.25);
                 moveRobot(speedX, 0, 0);
 
             // If the box is on the right of the robot, it moves right, and vise versa
